+++ examples/web.rs
@@ -32,7 +32,7 @@ pub fn main() -> Result<(), impl std::error::Error> {
                 event: WindowEvent::CloseRequested,
                 window_id,
             } if window_id == window.id() => control_flow.set_exit(),
-            Event::MainEventsCleared => {
+            Event::AboutToWait => {
                 window.request_redraw();
             }
             Event::WindowEvent {