+++ src/platform/pump_events.rs
@@ -74,7 +74,7 @@ pub trait EventLoopExtPumpEvents {
     ///                     event: WindowEvent::CloseRequested,
     ///                     window_id,
     ///                 } if window_id == window.id() => control_flow.set_exit(),
-    ///                 Event::MainEventsCleared => {
+    ///                 Event::AboutToWait => {
     ///                     window.request_redraw();
     ///                 }
     ///                 _ => (),